requirements.txt
streamlit
gitpython
langchain
langchain-community
langchain-groq
langchain-text-splitters
chromadb
sentence-transformers